mcp-server/pyproject.toml
[project]
name = "subscription-tracker-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
subscription-mcp = "server:main"
